Simplified the code a lot. Put back the denoiser hooks for further cancellation (changed spectral param to float* at least for now).
git-svn-id: http://svn.xiph.org/trunk/speex@9236 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ include/speex/speex_preprocess.h
@@ -109,10 +109,10 @@ SpeexPreprocessState *speex_preprocess_state_init(int frame_size, int sampling_r
 void speex_preprocess_state_destroy(SpeexPreprocessState *st);
 
 /** Preprocess a frame */
-int speex_preprocess(SpeexPreprocessState *st, spx_int16_t *x, int *echo);
+int speex_preprocess(SpeexPreprocessState *st, spx_int16_t *x, float *echo);
 
 /** Preprocess a frame */
-void speex_preprocess_estimate_update(SpeexPreprocessState *st, spx_int16_t *x, int *echo);
+void speex_preprocess_estimate_update(SpeexPreprocessState *st, spx_int16_t *x, float *echo);
 
 /** Used like the ioctl function to control the preprocessor parameters */
 int speex_preprocess_ctl(SpeexPreprocessState *st, int request, void *ptr);
